Update email query case sensitivity and bump version
The SQL query in the get_author_token_from_email method in afp.py has been updated to be case-insensitive when comparing emails. Previously, the search for emails was case-sensitive which could potentially miss some entries. The version in setup.py is then updated to reflect this change in the codebase.