Remove deprecated save/load corpus methods, refine existing tests
This commit removes the save and load methods in `wbtools/literature/corpus.py` as these methods were deprecated and no longer used in the codebase. They were initially used for storing and retrieving the analysed corpus to and from a file, which is no longer necessary with the current setup.

Moreover, several tests in `tests/literature/test_paper.py` and `tests/literature/test_corpus.py` were cleaned and updated to fit the present functionalities of the codebase.

This commit streamlines the codebase, removes legacy code and ensures the tests cover existing functionalities. This makes the project cleaner, more maintainable, and easier to expand in the future.